add useEditor util
